requirements.txt
torch>=1.13.0
accelerate>=0.20.0
diffusers>=0.21.0
pandas>=1.5.0
scikit-learn>=1.2.0
torchvision>=0.14.0
Pillow>=9.0.0
xformers>=0.0.20
matplotlib>=3.6.0
tqdm>=4.64.0
numpy>=1.21.0
safetensors>=0.3.0
